% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{PPT.AddRectangle}
\alias{PPT.AddRectangle}
\title{Add Rectangle shape}
\usage{
PPT.AddRectangle(ppt, top = 0.05, left = 0.05, width = 0.9,
  height = 0.9, fill.color = "grey", fill.transparency = 0,
  line.color = "black", line.type = 1, line.size = 1, maxscale = 1,
  newslide = FALSE)
}
\arguments{
\item{ppt}{The ppt object as used in \pkg{R2PPT}.}

\item{top, left}{Vertical and horizontal placement of the shape. Either as
fraction of slides dimensions or as pixel value. Values bigger than
\code{maxscale} are interpreted as pixels.#'}

\item{width, height}{Dimensions of shape. For values smaller than
\code{maxscale} (default is \code{1}) this refers to a proportion of the
current slide's width or height. Values bigger than \code{maxscale} are
interpreted as pixels.}

\item{fill.color}{Background color either as hex value or R color name.}

\item{fill.transparency}{Transparency of filling (\code{[0,1]}, 
default is \code{0} = opaque.).}

\item{line.color}{Color of text either as hex value or color name.}

\item{line.type}{\code{1} = solid (default), \code{2-8}= dots, dashes and
mixtures. See MsoLineDashStyle Enumeration for details.}

\item{line.size}{Thickness of line (default\code{1}).}

\item{maxscale}{Threshold below which values are interpreted as proportional
scaling factors for the \code{width} and \code{height} argument. Above the
threshold values are interpreted as pixels.}

\item{newslide}{Logical (default is \code{TRUE}) Whether the graphic will be
placed on a new slide.}
}
\description{
Add a rectangle to a slide. YOu can position it and modify a limited number
of aspects of its appearance (color etc.)
}
\examples{
\dontrun{
  
# add rectangle shapes to slide
p <- PPT.Init(visible=T, method = "RDCOMClient")
p <- PPT.AddBlankSlide(p)
p <- PPT.AddRectangle(p, height=.4)
p <- PPT.AddRectangle(p, top = .5, width = .4, height=.45, 
                      fill.color="blue", fill.transparency = .8) 
p <- PPT.AddRectangle(p, top = .5, width = .4, height=.45, left=.55, 
                      line.type=4, line.color="red", line.size = 3) 

}

}
\author{
Mark Heckmann
}
