% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/text.R
\name{PPT.ReplaceTextByText}
\alias{PPT.ReplaceTextByText}
\title{Replace matching text by text}
\usage{
PPT.ReplaceTextByText(ppt, what, replace, ...)
}
\arguments{
\item{ppt}{The ppt object as used in \pkg{R2PPT}.}

\item{what}{Text pattern to match against.}

\item{replace}{Text to replace pattern with.}

\item{...}{Not evaluated.}
}
\description{
Looks through all shapes and finds a shape with matching text pattern.
The text in the shape is replaced.
}
\examples{
\dontrun{
  
  # replace text on slide
  ppt <- PPT.Open("inst/template.pptx", method="RDCOMClient")

  what <- "[[tag 1]]"
  replace <- "This has been replaced"
  PPT.ReplaceTextByText(ppt, what, replace)
}
}

