% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{PPT.ShapesOnCurrentSlide}
\alias{PPT.ShapesOnCurrentSlide}
\title{Get pointers to all shapes on current slide}
\usage{
PPT.ShapesOnCurrentSlide(ppt)
}
\arguments{
\item{ppt}{The ppt object as used in \pkg{R2PPT}.}
}
\value{
List of pointers to shapes
}
\description{
Get pointers to all shapes on current slide
}
