% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text.R
\name{PPT.UpdateAutosizedTextboxes}
\alias{PPT.UpdateAutosizedTextboxes}
\title{Autosize textboxes}
\usage{
PPT.UpdateAutosizedTextboxes(ppt, slide_index = NULL)
}
\arguments{
\item{ppt}{The ppt object as used in \pkg{R2PPT}.}

\item{slide_index}{Slides indexes for which to update textboxes (default \code{NULL} = all slides).}
}
\description{
Sometimes text boxes which are set to autosize are not properly resized
when the file is opened. The autosizing is triggered by the function.
}
