% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{PPT.UpdateCurrentSlide}
\alias{PPT.UpdateCurrentSlide}
\title{Update the current slide stored in R2PPT object}
\usage{
PPT.UpdateCurrentSlide(ppt, i = NULL, slide = NULL)
}
\arguments{
\item{ppt}{The ppt object as used in \pkg{R2PPT}.}

\item{i}{Slide index.}

\item{slide}{A slide object as alternative to setting the index.}
}
\description{
R2PPT uses an object to store the current slide amongst other things. 
Unfortunately the current slide is only set when a new slide is inserted. It
is NULL when a file is loaded. This will cause errors sometimes, hence we may
need to set it manually.
}
\author{
Mark Heckmann
}
