% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/R2PPTaddons.R
\docType{package}
\name{R2PPTaddons}
\alias{R2PPTaddons}
\title{\pkg{R2PPTaddons} - Feature enhancement for \pkg{R2PPT}.}
\description{
Functions to be used for creating output in MS PowerPoint.
Builds upon \pkg{R2PPT}.
}
\author{
Mark Heckmann
}
