% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table.R
\name{PPT.AddTable}
\alias{PPT.AddTable}
\title{Adding a table to a slide.}
\usage{
PPT.AddTable(ppt, df, width = 0.9, height = 0.9, cell.height = 0.01,
  col.align = "left", left = 0.05, top = 0.1, colnames = TRUE,
  rownames = FALSE, newslide = FALSE, maxscale = 1)
}
\description{
Adding a table to a slide.
}
\examples{
\dontrun{

# open new PPT presentation
p <- PPT.Init(visible=T, method = "RDCOMClient")  

## EXAMPLE 1 ##

p <- PPT.AddBlankSlide(p)
m <- mtcars[1:2, ]
p <- PPT.AddTable(p, m, colnames = F, rownames = F, top=.1)
p <- PPT.AddTable(p, m, colnames = T, rownames = F, top=.25)
p <- PPT.AddTable(p, m, colnames = F, rownames = T, top=.45)
p <- PPT.AddTable(p, m, colnames = T, rownames = T, top=.7)

}

}
