% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shapes.R
\name{PPT.AddTextBox}
\alias{PPT.AddTextBox}
\title{Adding a textbox to a slide.}
\usage{
PPT.AddTextBox(ppt, txt, width = 0.9, height = 0.9, x = "center",
  y = "center", xy.format = "points", x.offset = 0, y.offset = 0,
  offset.format = "points", x.text.align = "center",
  bullet.points = "none", bullet.points.color = 0, font.size = 16,
  font.color = "black", font.bold = FALSE, font.italic = FALSE,
  fill.color = NA, border.color = NA, newslide = FALSE, maxscale = 1,
  autosize = TRUE, zcmd = 0, zposition = NULL)
}
\arguments{
\item{ppt}{The ppt object as used in \pkg{R2PPT}.}

\item{txt}{Text to put into the textbox. A vector with length greater 1
is collapsed using linebreak characters.}

\item{width}{Width of graphic. For values smaller than \code{maxscale}
(default is \code{1}) this refers to a proportion of the current slide's
width. Values bigger than \code{maxscale} are interpreted as pixels.If
\code{NA} only the height argument is used for sclaing.}

\item{height}{Height of graphic. For values smaller than \code{maxscale}
(default is \code{1}) this refers to a proportion of the current slide's
height. Values bigger than \code{maxscale} are interpreted as pixels. If
\code{NA} only the width argument is used for scaling.}

\item{x}{Horizontal placement of the textbox. Either a string
(\code{"center", "left", "right"}) or a numerical value. Numerical values
are interpreted as absolute position in pixels counted from the left of the
slide.}

\item{y}{Vertical placenment of the textbox. Either a string (\code{"center",
"top", "bottom"}) or a numerical value. Numerical values are interpreted as
absolute position in pixels counted from the top of the slide.}

\item{xy.format}{The numeric x and y input will be interpreted either as
\code{"pixels"} (default) or \code{"percent"} of the slide's total
width/height. Character input will not be affected.}

\item{x.offset}{Additional horizontal offset in pixel. Used for finetuning
position on slide.}

\item{y.offset}{Additional horizontal offset in pixel or as percent (see .
Used for finetuning position on slide.}

\item{offset.format}{The offset will be interpreted either as \code{"pixels"}
(default) or \code{"percent"} of the slide's total width/height.}

\item{x.text.align}{Horizontal alignment of text (\code{"left", "center",
"right"}).}

\item{bullet.points}{Whether to treat each new line and vector element as a
bullet point (\code{"none", "unnumbered", "numbered"}).}

\item{bullet.points.color}{Color of bullet points either as hex value or
color name.}

\item{font.size}{Text size (default 16).}

\item{font.color}{Color of text either as hex value or color name.}

\item{font.bold}{Bold text (default \code{FALSE}).}

\item{font.italic}{Italic text (default \code{FALSE}).}

\item{fill.color}{Background color either as hex value or color name.}

\item{border.color}{Border line color either as hex value or color name.}

\item{newslide}{Logical (default is \code{TRUE}) Whether the graphic will be
placed on a new slide.}

\item{maxscale}{Threshold below which values are interpreted as proportional
scaling factors for the \code{width} and \code{height} argument. Above the
threshold values are interpreted as pixels.}

\item{zcmd}{Determine a change in z-order after shape is added. 
Either a number or a text abbreviation (e.g. \code{"front"}). 
See \code{\link{set_shape_zorder}} for more details.}

\item{zposition}{Set the z-position directly using a numeric. A higher z-order means on top of other shapes.
Numbers outside the possible range (no of shapes) is fixed to allowed range.}
}
\description{
Add a textbox to a slide. YOu can easily position it and modify a limited
number of aspects of its appearance (color, bullet points, etc.)
}
\examples{
\dontrun{
  
# add textbox to middle of slide
p <- PPT.Init(visible=T, method = "RDCOMClient")
p <- PPT.AddBlankSlide(p)
txt = c("Line 1", "Line 2", "Line 3")
p <- PPT.AddTextBox(p, txt) 

}

}
\author{
Mark Heckmann
}
