% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graphics.R
\name{PPT.ReplaceTextByGraphic}
\alias{PPT.ReplaceTextByGraphic}
\title{Replace matching text by graphic}
\usage{
PPT.ReplaceTextByGraphic(ppt, what, file, shape.type = 17, ...)
}
\arguments{
\item{ppt}{The ppt object as used in \pkg{R2PPT}.}

\item{what}{Text pattern to match against.}

\item{file}{Path to the graphic file.}

\item{shape.type}{Shape types in which the text pattern is searched for. By
default only plain text fields (\code{shape.type = 17}) are searched. Other
shapes, e.g. rectangles with text, are ignored. To search all shapes use
\code{shape.type = NA}. The types are documented in the
\code{MsoAutoShapeType} enumeration in Microsoft's MSDN docu.}

\item{...}{Arguments passed on to \code{\link{PPT.AddGraphicstoSlide2}}.}
}
\description{
Looks through all shapes and finds a shape with matching text pattern. The
shape is deleted and a graphic is inserted on the shape's parent slide.
}
\examples{
\dontrun{

# open PPT template
file <- system.file("template.pptx", package = "R2PPTaddons")
ppt <- PPT.Open(file, method="RDCOMClient")


#### CASE 1: ADD GRAPHIC AT TEXTBOX ####

# add graphic to slides with matching text and remove text. Note that by default
# only text in text boxes is replaced. Text inside rectangles, for example, is
# not touched unless explicitly prompted. Note also that additional arguments
# are passed on to PPT.AddGraphicstoSlide2 to specify the position and size of
# the shape.
what <- "[[tag 1]]"
PPT.ReplaceTextByGraphic(ppt, what, file = "inst/image_1.png", width = .6)

# Note that the text appears twice and the graphic is inserted twice with a warning.
what <- "[[tag 2]]"
PPT.ReplaceTextByGraphic(ppt, what, file = "inst/image_2.png")



#### CASE 2: REPLACE SHAPE BY GRAPHIC ####

# replace rectangle with matching text by graphic and move to back
what <- "[[tag 1]]"
PPT.ReplaceShapeByGraphic(ppt, what, file = "inst/image_1.png", z.order = 1)

# Note  that additional arguments are passed on to PPT.FitGraphicIntoShape and 
# subsequently to PPT.AddGraphicstoSlide2, e.g. to specify the appearance of the image.
what <- "[[tag 2]]"
PPT.ReplaceShapeByGraphic(ppt, what, file = "inst/image_2.png", line.size=1)

}




}
\seealso{
PPT.ReplaceShapeByGraphic
}
\author{
Mark Heckmann
}
