% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/graphics.R
\name{PPT.ReplaceTextByGraphic}
\alias{PPT.ReplaceTextByGraphic}
\title{Replace matching text by graphic}
\usage{
PPT.ReplaceTextByGraphic(ppt, what, file, ...)
}
\arguments{
\item{ppt}{The ppt object as used in \pkg{R2PPT}.}

\item{what}{Text pattern to match against.}

\item{file}{Path to the graphic file.}

\item{...}{Arguments passed on to \code{\link{PPT.AddGraphicstoSlide2}}.}
}
\description{
Looks through all shapes and finds a shape with matching text pattern.
The shape is deleted and a graphic is inserted on the shape's parent slide.
}
\examples{
\dontrun{
  
# add graphic to slides with matching and remove text

ppt <- PPT.Open("inst/template.pptx", method="RDCOMClient")

what <- "[[tag 1]]"
PPT.ReplaceTextByGraphic(ppt, what, file = "inst/image_1.png")

# Note that the text appears twice and the graphic is inserted twice with a warning.
what <- "[[tag 2]]"
PPT.ReplaceTextByGraphic(ppt, what, file = "inst/image_2.png")

}

}
\author{
Mark Heckmann
}

